Using float version of log instead of double version shaves a little bit off execution for simple models.
git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/branches/misc_opts@14600 9cb281f4-553f-406f-8d7f-11f072bbfe46

+++ src/MCMC.c
@@ -267,7 +267,7 @@ MCMCStatus MetropolisHastings(MHproposal *MHp,
     }
     
     /* if we accept the proposed network */
-    if (cutoff >= 0.0 || log(unif_rand()) < cutoff) { 
+    if (cutoff >= 0.0 || logf(unif_rand()) < cutoff) { 
       if(fVerbose>=5){
 	Rprintf("Accepted.\n");
       }
